* commit '1109683752f3f3d57b636fa49e719c8e01103573': (34 commits)
  COP-14548: Fixed comments.
  COP-14548: Refactored removeProtection to removeProtectionStep and added return type as boolean.
  COP-14548: Spelling fix.
  COP-14548: Grammar fix.
  COP-14548: Moved if up into else if.
  COP-14548: Updated javadoc.
  COP-14548: Grammar fix
  COP-14548: Added more logging constants
  COP-14548: Update log sev to WARN.
  COP-14548: Removed colon.
  COP-14548: Ensure that vplex export steps do not get fired if the do not delete flag is set on the VPLEX volume.
  COP-14548: Fixes for VPLEX do not delete volume code. Also added common method to get the volumes to not delete from the descriptors.
  COP-14548: Update RP sanity for change vpool tests.
  COP-14548: Added consistencyGroup tag to change vpool test.
  COP-14548: Made reusable log strings as constants in RPHelper. Small fix in rp sanity.
  COP-14548: Null checks for decriptor param objects and added more comments.
  COP-14548: Moved the dependency for cleaning up the VPLEX volumes from their CGs to the VPLEXDeviceController.
  COP-17482: set missing coordinator reference on CapacityMatcher
  COP-14548: Update sanity for RP remove protection. Also add CG param to volume cos change.
  COP-14548: Minor logging corrections.
  ...

+++ controllersvc/src/main/java/com/emc/storageos/vplexcontroller/VPlexDeviceController.java
@@ -970,7 +970,13 @@ public class VPlexDeviceController implements VPlexController, BlockOrchestratio
             if (vplexVolumes.isEmpty()) {
                 return waitFor;
             }
-
+            
+            // Check to see if there are any volumes flagged to not be fully deleted.
+            // This will still remove the volume from its VPLEX CG and also clean up
+            // any Mirrors but will leave the Virtual Volume intact on the VPLEX.
+            List<VolumeDescriptor> doNotDeleteDescriptors = VolumeDescriptor.getDoNotDeleteDescriptors(vplexVolumes);
+            List<URI> doNotFullyDeleteVolumeList = VolumeDescriptor.getVolumeURIs(doNotDeleteDescriptors);
+            
             List<URI> allVplexVolumeURIs = VolumeDescriptor.getVolumeURIs(vplexVolumes);
 
             // Segregate by device.
@@ -984,7 +990,7 @@ public class VPlexDeviceController implements VPlexController, BlockOrchestratio
                                 BlockDeviceController.getVolumesMsg(_dbClient, vplexVolumeURIs)),
                         waitFor, vplexURI,
                         DiscoveredDataObject.Type.vplex.name(), this.getClass(),
-                        deleteVirtualVolumesMethod(vplexURI, vplexVolumeURIs),
+                        deleteVirtualVolumesMethod(vplexURI, vplexVolumeURIs, doNotFullyDeleteVolumeList),
                         rollbackMethodNullMethod(), null);
             }
 
@@ -1000,8 +1006,10 @@ public class VPlexDeviceController implements VPlexController, BlockOrchestratio
             List<URI> backendVolURIs = new ArrayList<URI>();
             for (URI vplexVolumeURI : allVplexVolumeURIs) {
                 Volume vplexVolume = _dbClient.queryObject(Volume.class, vplexVolumeURI);
-                if ((vplexVolume == null) || (vplexVolume.getInactive()) ||
-                        (vplexVolume.isIngestedVolume(_dbClient))) {
+                if ((vplexVolume == null) 
+                        || (vplexVolume.getInactive()) 
+                        || (vplexVolume.isIngestedVolume(_dbClient))
+                        || doNotFullyDeleteVolumeList.contains(vplexVolumeURI)) {
                     continue;
                 }
                 for (String assocVolumeId : vplexVolume.getAssociatedVolumes()) {
@@ -1120,16 +1128,22 @@ public class VPlexDeviceController implements VPlexController, BlockOrchestratio
         // Filter to get only the VPlex volumes.
         List<VolumeDescriptor> vplexVolumes = VolumeDescriptor.filterByType(volumes,
                 new VolumeDescriptor.Type[] { VolumeDescriptor.Type.VPLEX_VIRT_VOLUME },
-                new VolumeDescriptor.Type[] {});
+                new VolumeDescriptor.Type[] {});        
+        // Check to see if there are any volumes flagged to not be fully deleted.
+        // Any flagged volumes will be removed from the list of volumes to delete.
+        List<VolumeDescriptor> descriptorsToRemove = VolumeDescriptor.getDoNotDeleteDescriptors(vplexVolumes);
+        vplexVolumes.removeAll(descriptorsToRemove);
+        
         // If there are no VPlex volumes, just return
         if (vplexVolumes.isEmpty()) {
             return waitFor;
         }
+        
         URI vplexURI = vplexVolumes.get(0).getDeviceURI();
-
+        
         // Get the VPlex Volume URIs
         List<URI> allVplexVolumeURIs = VolumeDescriptor.getVolumeURIs(vplexVolumes);
-
+                
         // Add a step to the Workflow to mark the Virtual Volumes inactive.
         // Rollback does the same thing.
         waitFor = workflow.createStep(null, "Mark virtual volumes inactive", waitFor,
@@ -1300,9 +1314,9 @@ public class VPlexDeviceController implements VPlexController, BlockOrchestratio
         WorkflowStepCompleter.stepSucceded(stepId);
     }
 
-    private Workflow.Method deleteVirtualVolumesMethod(URI vplexURI, List<URI> volumeURIs) {
+    private Workflow.Method deleteVirtualVolumesMethod(URI vplexURI, List<URI> volumeURIs, List<URI> doNotFullyDeleteVolumeList) {
         return new Workflow.Method(DELETE_VIRTUAL_VOLUMES_METHOD_NAME, vplexURI,
-                volumeURIs);
+                volumeURIs, doNotFullyDeleteVolumeList);
     }
 
     /**
@@ -1312,10 +1326,12 @@ public class VPlexDeviceController implements VPlexController, BlockOrchestratio
      * 
      * @param vplexURI
      * @param volumeURIs
+     * @param doNotFullyDeleteVolumeList
      * @param stepId
      * @throws WorkflowException
      */
-    public void deleteVirtualVolumes(URI vplexURI, List<URI> volumeURIs, String stepId) throws WorkflowException {
+    public void deleteVirtualVolumes(URI vplexURI, List<URI> volumeURIs, 
+            List<URI> doNotFullyDeleteVolumeList, String stepId) throws WorkflowException {
         try {
             WorkflowStepCompleter.stepExecuting(stepId);
             VPlexApiClient client = getVPlexAPIClient(_vplexApiFactory, vplexURI, _dbClient);
@@ -1335,9 +1351,7 @@ public class VPlexDeviceController implements VPlexController, BlockOrchestratio
                     continue;
                 }
                 try {
-                    client.findVirtualVolume(volume.getDeviceLabel());
-                    _log.info(String.format("Deleting VPlex virutal volume %s (%s)",
-                            volume.getDeviceLabel(), volume.getNativeId()));
+                    client.findVirtualVolume(volume.getDeviceLabel());                    
                 } catch (VPlexApiException ex) {
                     if (ex.getServiceCode() == ServiceCode.VPLEX_CANT_FIND_REQUESTED_VOLUME) {
                         _log.info("VPlex virtual volume: " + volume.getNativeId()
@@ -1361,12 +1375,21 @@ public class VPlexDeviceController implements VPlexController, BlockOrchestratio
                             consistencyGroupManager.deleteConsistencyGroupVolume(vplexURI, volume, cg.getLabel());
                         }
 
-                        // We only retry a dismantle failure on volumes created
-                        // in ViPR as the retry code relies on the well-known ViPR
-                        // naming conventions and virtual volume structure to find
-                        // VPLEX artifacts related to the volume being deleted.
-                        boolean isIngested = volume.isIngestedVolume(_dbClient);
-                        client.deleteVirtualVolume(volume.getDeviceLabel(), !isIngested, !isIngested);
+                        // Check to see if there are any entries in the doNotFullyDeleteVolumeList.
+                        // If this volume ID is flagged (in that list) we can skip the call
+                        // to delete volume.
+                        if (doNotFullyDeleteVolumeList == null
+                                || doNotFullyDeleteVolumeList.isEmpty() 
+                                || !doNotFullyDeleteVolumeList.contains(volume.getId())) {
+                            // We only retry a dismantle failure on volumes created
+                            // in ViPR as the retry code relies on the well-known ViPR
+                            // naming conventions and virtual volume structure to find
+                            // VPLEX artifacts related to the volume being deleted.
+                            _log.info(String.format("Deleting VPlex virtual volume %s (%s)",
+                                    volume.getDeviceLabel(), volume.getNativeId()));
+                            boolean isIngested = volume.isIngestedVolume(_dbClient);
+                            client.deleteVirtualVolume(volume.getDeviceLabel(), !isIngested, !isIngested);
+                        }
 
                         // Record VPLEX volume deleted event.
                         recordBourneVolumeEvent(volume.getId(),
@@ -5644,7 +5667,7 @@ public class VPlexDeviceController implements VPlexController, BlockOrchestratio
                 // If importing to a local/distributed virtual volume, then
                 // rollback the VPlex create virtual volumes operation.
                 // We will restore the original volume.
-                vplexRollbackMethod = deleteVirtualVolumesMethod(vplexURI, vplexVolumeURIs);
+                vplexRollbackMethod = deleteVirtualVolumesMethod(vplexURI, vplexVolumeURIs, null);
             } else {
                 // COP-16861: If rolling back an upgrade from local to distributed, then
                 // try to detach remote mirror and delete new artifacts created on VPLEX
